Clean up code
