+++ action_interface.rb
@@ -1,0 +1,9 @@
+class ActionInterface
+  def create
+    raise NotImplementedError, 'Subclasses must implement this method'
+  end
+
+  def list_all
+    raise NotImplementedError, 'Subclasses must implement this method'
+  end
+end
